Remove dead user code and simplify to renderables
This also breaks renderables and parts out for users from both markup
and styles sides.
