--- src/components/users/UserCard.css
@@ -1,52 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-@media (--break-2) {
-  .Cards {
-    margin-left: -20px;
-  }
-
-  .Cards .UserCard {
-    display: inline-block;
-    width: calc(50% - 20px);
-    height: 320px;
-    margin-top: 20px;
-    margin-left: 20px;
-  }
-}
-
-@media (--break-3) {
-  .Cards .UserCard {
-    width: calc(33.33333% - 20px);
-  }
-}
-
-@media (--break-4) {
-  .Cards {
-    margin-left: -40px;
-  }
-
-  .Cards .UserCard {
-    width: calc(25% - 40px);
-    margin-top: 40px;
-    margin-left: 40px;
-  }
-}
-
-@media (--break-5) {
-  .Cards .UserCard {
-    width: calc(20% - 40px);
-  }
-}
-
-@media (--break-6) {
-  .Cards .UserCard {
-    width: calc(16.66666% - 40px);
-  }
-}
-
-@media (--break-7) {
-  .Cards .UserCard {
-    width: calc(14.28571% - 40px);
-  }
-}
-
