--- src/components/users/UserCompact.css
@@ -1,41 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-.UserCompact {
-  position: relative;
-  height: 40px;
-  margin-bottom: 40px;
-  overflow: hidden;
-  font-size: 14px;
-}
-
-.UserCompactHeader {
-  display: inline-block;
-  width: calc(100% - 120px);
-  height: 40px;
-  overflow: hidden;
-  line-height: 40px;
-}
-
-.UserCompactUserLink {
-  display: inline-block;
-  transition: color var(--speed) ease;
-}
-
-.no-touch .UserCompactUserLink:hover {
-  color: #999;
-}
-
-.UserCompactUsername {
-  padding-left: 10px;
-}
-
-.UserCompact > .RelationshipContainer {
-  position: absolute;
-  top: 5px;
-  right: -8px;
-}
-
-.UserCompact > .RelationshipContainer .StarshipButton {
-  color: #aaa;
-}
-
