--- src/components/users/UserDrawer.css
@@ -1,26 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-.UserDrawer {
-  position: relative;
-  max-width: 780px;
-  min-height: 90px;
-  padding-left: 20px;
-}
-
-.UserDrawer > .SVGIcon {
-  position: absolute;
-  top: 15px;
-  left: 0;
-}
-
-.UserDrawer .UserAvatar {
-  margin-right: 10px;
-  margin-bottom: 10px;
-}
-
-@media (--break-2) {
-  .UserDrawer > .SVGIcon {
-    top: 30px;
-  }
-}
-
