--- src/components/users/UserIcons.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import { SVGIcon } from '../svg/SVGComponents'
-
-export const ShareIcon = () =>
-  <SVGIcon className="ShareIcon">
-    <g>
-      <polyline points="7.8,7.3 5,7.3 5,17.3 15,17.3 15,7.3 12.2,7.3" />
-    </g>
-    <g>
-      <line x1="10" y1="2" x2="10" y2="12" />
-      <polyline points="7.2,4.7 10,2 12.8,4.8" />
-    </g>
-  </SVGIcon>
-
-export default ShareIcon
-
