+++ src/components/users/UserParts.css
@@ -2,101 +2,8 @@
 
 /* stylelint-disable no-unsupported-browser-features */
 
-/* -------------------------------------
-Users Collections  */
-
-@media (--break-2) {
-  .Users.asGrid {
-    margin-left: -20px;
-  }
-}
-
-@media (--break-4) {
-  .Users.asGrid {
-    margin-left: -40px;
-  }
-}
-
-/* -------------------------------------
-User Profile  */
-
-.UserProfile {
-  position: relative;
-  flex-grow: 1;
-  max-width: 600px;
-  padding-right: 10px;
-  padding-left: 10px;
-  margin: auto;
-  color: #000;
-  text-align: center;
-}
-
-@media (--break-2) {
-  .UserProfile {
-    min-height: 360px;
-    padding-right: 0;
-    padding-left: 0;
-    color: #fff;
-  }
-}
-
-/* -------------------------------------
-User Cards  */
-
-.UserProfileCard {
-  position: relative;
-  height: 250px;
-  padding: 20px;
-  margin-top: 10px;
-  overflow: hidden;
-  color: #fff;
-  text-align: center;
-  background-color: transparent;
-  border-radius: 5px;
-}
-
-@media (--break-2) {
-  .UserProfileCard {
-    display: inline-block;
-    width: calc(50% - 20px);
-    height: 540px;
-    margin-top: 20px;
-    margin-left: 20px;
-    vertical-align: top;
-  }
-}
-
-@media (--break-3) {
-  .UserProfileCard {
-    width: calc(33.33333% - 20px);
-  }
-}
-
-@media (--break-4) {
-  .UserProfileCard {
-    width: calc(25% - 40px);
-    margin-top: 40px;
-    margin-left: 40px;
-  }
-}
-
-@media (--break-5) {
-  .UserProfileCard {
-    width: calc(20% - 40px);
-  }
-}
-
-@media (--break-6) {
-  .UserProfileCard {
-    width: calc(16.66666% - 40px);
-  }
-}
-
-@media (--break-7) {
-  .UserProfileCard {
-    width: calc(14.28571% - 40px);
-  }
-}
+/* These are the basic building blocks for putting together
+ * UserRenderables */
 
 /* -------------------------------------
 User Cells  */
