+++ test/unit/actions/discover_test.js
@@ -4,7 +4,6 @@ import { mostRecentPostsFromUsers } from '../../../src/components/streams/Stream
 import {
   postsAsGrid,
   postsAsList,
-  usersAsCards,
   usersAsGrid,
 } from '../../../src/components/streams/StreamRenderables'
 
@@ -84,8 +83,8 @@ describe('discover actions', () => {
     })
 
     it('has asList, asGrid and asZero properties on renderStreams in the action', () => {
-      expect(action.meta.renderStream.asList).to.equal(usersAsCards)
-      expect(action.meta.renderStream.asGrid).to.equal(usersAsCards)
+      expect(action.meta.renderStream.asList).to.equal(usersAsGrid)
+      expect(action.meta.renderStream.asGrid).to.equal(usersAsGrid)
     })
   })
